# backend/swagger.yaml
openapi: 3.0.0
info:
  title: E-commerce API
  description: API documentation for the E-commerce backend.
  version: 3.0.0
servers:
  - url: http://localhost:3000
    description: Local development server

#  user api documentation
paths:
...
  /category/list:
    get:
      summary: Get all categories
      description: Retrieve a list of all category.
      tags:
        - Category
      operationId: categoryList
      responses:
        "200":
          description: Successfully retrieved category list
          content:
            application/json:
              schema:
                type: object
                properties:
                  categoryList:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: The unique identifier of the category.
                        name:
                          type: string
                          description: The name of the category.
                        status:
                          type: string
                          description: The status of the category .
                        isDeleted:
                          type: boolean
                          description: Whether the category has been deleted .
        "500":
          description: Internal server error

  # category delete api documentation
...
  /subCategory/update/{id}:
    put:
      summary: Admin  update a subCategory
      description: update subCategory
      tags:
        - subCategory
      operationId: updateSubCategory
      parameters:
        - name: id
          in: path
          required: true
          description: subCategory Id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryId:
                  type: string
                  description: Category Id
                name:
                  type: string
                  description: subCategory name
              required:
                - categoryId
                - name
      responses:
        "200":
          description: subCategory updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  name:
                    type: string
        "500":
          description: Internal server error

  #subCategory delete api documentation
  /subCategory/delete/{id}:
    delete:
      summary: Admin delete subCategory
      tags:
        - subCategory
      operationId: deleteSubCategory
      parameters:
        - name: id
          in: path
          required: true
          description: subCategory Id
          schema:
            type: string
      responses:
        "200":
          description: subCategory deleted successfully
        "500":
          description: Internal server error

  # subCategory list get api documentation
...
  /review/delete/{id}:
    delete:
      summary: Deleted review
      description: Delete review by review Id
      tags:
        - Review
      operationId: deleteReview
      parameters:
        - name: id
          in: path
          required: true
          description: review Id
          schema:
          type: string
      responses:
        "200":
          description: review deleted successfully
        "500":
          description: Internal server error

  #cart add post api documentation
...
  /wishlist/delete/{id}:
    delete:
      summary: wishlist deleted
      tags:
        - Wishlist
      operationId: deleteWishlist
      parameters:
        - name: id
          in: path
          required: true
          description: wishlist Id
          schema:
            type: string
      responses:
        "200":
          description: Wishlist item successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: Product have removed successfully
        "500":
          description: Internal server error

  #discount  post api documentation
...